14 Task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,18 +6,18 @@
  */
 int main(void)
 {
-	int x, a, b, c;
+	long x, a, b, c;
 
 	a = 0;
 	b = 1;
 	c = a + b;
 	for (x = 1; x < 50; x++)
 	{
-		printf("%d, ", c);
+		printf("%ld, ", c);
 		a = b;
 		b = c;
 		c = a + b;
 	}
-	printf("%d\n", c);
+	printf("%ld\n", c);
 	return (0);
 }
